[refector(Search,SearchDetail)] groupItem 함수 mapper클래스로 이전
